--- feature-ui-home/src/main/kotlin/team/duckie/app/android/feature/ui/home/empty.kt
@@ -1,8 +1,0 @@
-/*
- * Designed and developed by Duckie Team, 2022
- *
- * Licensed under the MIT.
- * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
- */
-
-package team.duckie.app.android.feature.ui.home
