+++ feature-ui-search/build.gradle.kts
@@ -10,10 +10,7 @@ import DependencyHandler.Extensions.implementations
 plugins {
     id(ConventionEnum.AndroidLibrary)
     id(ConventionEnum.AndroidLibraryCompose)
-    id(ConventionEnum.AndroidLibraryComposeUiTest)
     id(ConventionEnum.AndroidHilt)
-    id(ConventionEnum.JvmJUnit4)
-    id(ConventionEnum.JvmDokka)
 }
 
 android {
@@ -32,7 +29,6 @@ dependencies {
         projects.sharedUiCompose,
         libs.quack.ui.components,
         libs.compose.ktx.lifecycle,
-        libs.compose.ui.coil,
         libs.firebase.crashlytics,
     )
 }
