+++ feature-ui-search/src/main/AndroidManifest.xml
@@ -1,5 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
+<?xml version="1.0" encoding="utf-8"?><!--
   ~ Designed and developed by Duckie Team, 2022
   ~
   ~ Licensed under the MIT.
