+++ feature-ui-search/src/main/kotlin/team/duckie/app/android/feature/ui/search/viewmodel/sideeffect/SearchResultSideEffect.kt
@@ -7,8 +7,7 @@
 
 package team.duckie.app.android.feature.ui.search.viewmodel.sideeffect
 
-sealed class SearchResultSideEffect {
-
+internal sealed class SearchResultSideEffect {
     /**
      * [SearchViewModel] 의 비즈니스 로직 처리 중에 발생한 예외를 [exception] 으로 받고
      * 해당 exception 을 [FirebaseCrashlytics] 에 제보합니다.
